requirements.txt
ag2==0.9.10
annotated-types==0.7.0
anyio==4.11.0
appscript==1.3.0 ; sys_platform == 'darwin'
autogen-agentchat==0.7.5
autogen-core==0.7.5
certifi==2025.8.3
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==46.0.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docx2pdf==0.1.8
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
fpdf==1.7.2
grandalf==0.8
greenlet==3.2.4 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain==0.3.27
langchain-core==0.3.78
langchain-openai==0.3.34
langchain-text-splitters==0.3.11
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.32
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
openai==2.1.0
opentelemetry-api==1.37.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==11.3.0
protobuf==5.29.5
pyautogen==0.10.0
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
pydantic==2.11.10
pydantic-core==2.33.2
pygments==2.19.2
pymupdf==1.26.4
pyparsing==3.2.5
pypdfium2==4.30.0
pytesseract==0.3.13
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311 ; sys_platform == 'win32'
pyyaml==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.41.0
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.11.0
tqdm==4.67.1
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
